Update DFU Blaster
This switches the existing DFU Blaster recipe to use the stable URL for the latest official release.

Previously it was pulling beta releases from BitBucket since that was all that existed when the recipe was originally created.

To continue downloading/importing beta releases switch to the new recipes that maintain the original logic and download source.